+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Added
 - Support for ADBS free DB version details
+- Support for MySQL Analytics added
 
 ## 4.2.0 (November 04, 2020)